Cache all query results by unique param sets for better API performance
