+++ app/src/main/kotlin/team/finder/api/App.kt
@@ -10,10 +10,12 @@ import io.micrometer.jmx.JmxConfig
 import io.micrometer.jmx.JmxMeterRegistry
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.cache.annotation.EnableCaching
 import org.springframework.context.annotation.Bean
 
 
 @SpringBootApplication
+@EnableCaching
 class App {
 
     @Bean
