+++ app/src/main/kotlin/team/finder/api/system/RedisConfig.kt
@@ -1,0 +1,15 @@
+package team.finder.api.system
+
+import org.springframework.context.annotation.Bean
+import org.springframework.data.redis.cache.RedisCacheConfiguration
+import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer
+import org.springframework.data.redis.serializer.RedisSerializationContext.SerializationPair
+import java.time.Duration
+
+@Bean
+fun cacheConfiguration(): RedisCacheConfiguration {
+    return RedisCacheConfiguration.defaultCacheConfig()
+        .entryTtl(Duration.ofMinutes(15))
+        .disableCachingNullValues()
+        .serializeValuesWith(SerializationPair.fromSerializer(GenericJackson2JsonRedisSerializer()))
+}
