+++ app/src/main/kotlin/team/finder/api/teams/Team.kt
@@ -1,13 +1,17 @@
 package team.finder.api.teams
 
+import kotlinx.serialization.Serializable
 import team.finder.api.utils.TimestampUtils
 import javax.persistence.Entity
 import javax.persistence.Id
 import javax.persistence.Table
+import java.io.Serializable as JavaSerializable
+
 
 @Entity
 @Table(name = "team")
-class Team(
+@Serializable
+class Team (
     var author: String?,
     var authorId: Long?,
     var description: String,
@@ -19,7 +23,7 @@ class Team(
     var deletedAt: String?,
 
     @Id val id: Long
-) {
+) : JavaSerializable {
 
     constructor(_author: String?, _authorId: Long?, _description: String, _skillsetMask: Int) :
             this(_author, _authorId, _description, _skillsetMask, TimestampUtils.getCurrentTimeStamp(), TimestampUtils.getCurrentTimeStamp(), null, 0)
