+++ app/src/main/kotlin/team/finder/api/teams/TeamsService.kt
@@ -1,5 +1,10 @@
 package team.finder.api.teams
 
+import io.micrometer.core.instrument.Counter
+import io.micrometer.core.instrument.Metrics
+import io.micrometer.core.instrument.Timer
+import org.springframework.cache.annotation.Cacheable
+import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
 import org.springframework.stereotype.Service
@@ -10,10 +15,41 @@ import java.util.*
 @Service
 class TeamsService(val repository: TeamsRepository) {
 
+    /**
+     * How many records should be returned per page?
+     */
+    val pageSize: Int = 25
+
+    val queryCounter: Counter = Metrics.counter("teams.counter")
+    val queryTimer: Timer = Metrics.timer("teams.query")
+
+
     fun createTeam(team: Team) = repository.save(team)
 
-    fun getTeams(pageable: Pageable): List<Team> = repository.getTeams(pageable)
-    fun getTeams(pageable: Pageable, skillsetMask: Int): List<Team> = repository.getTeams(pageable, skillsetMask)
+    @Cacheable("teams")
+    fun getTeams(pageIdx: Int, skillsetMask: Int, strSortingOption: String): List<Team> {
+        queryCounter.increment()
+
+        // A power-of-2 mask being set to 0 is meaningless - AKA "do not use"
+        val willPerformNativeQuery: Boolean = skillsetMask > 0
+        val sort: Sort = getSort(strSortingOption, willPerformNativeQuery)
+
+        // Pagination needs to be offset by -1 from expectations, but can't be set below 0
+        val queryPageable: PageRequest = PageRequest.of(pageIdx - 1, pageSize, sort)
+
+        var teams: List<Team> = listOf()
+
+
+        queryTimer.record {
+            teams = if (willPerformNativeQuery) {
+                repository.getTeams(queryPageable, skillsetMask)
+            } else {
+                repository.getTeams(queryPageable)
+            }
+        }
+
+        return teams
+    }
 
     fun getTeamByAuthorId(authorId: Long): Optional<Team> = repository.getTeamByAuthorId(authorId)
 
@@ -69,4 +105,5 @@ class TeamsService(val repository: TeamsRepository) {
             else        -> SortingOptions.Desc
         }
     }
+
 }
